refactor(validation.py): + moved make_check_date_between to a separate module
