+++ otri/validation/validation.py
@@ -1,7 +1,6 @@
 from typing import Sequence, Callable, Mapping, List, Any, Set, Tuple, Final
 from ..filtering.filter import Filter
 from ..filtering.stream import Stream
-from datetime import datetime
 
 
 class ValidatorFilter(Filter):
@@ -17,14 +16,6 @@ class ValidatorFilter(Filter):
     WARNING: Final = 2
     ERR_KEY: Final = "error"
     WARN_KEY: Final = "warning"
-
-    # Mapping methods to avoid if-else chain.
-    __ACTIONS = {
-        ValidatorFilter.OK: ValidatorFilter._on_ok,
-        ValidatorFilter.UNCLEAR: ValidatorFilter._on_unclear,
-        ValidatorFilter.ERROR: ValidatorFilter._on_error,
-        ValidatorFilter.WARNING: ValidatorFilter._on_warning
-    }
 
     def _on_data(self, data, index):
         '''
@@ -110,6 +101,15 @@ class ValidatorFilter(Filter):
             atom[key] = list()
         atom[key].append(value)
 
+    # Mapping methods to avoid if-else chain.
+    # Needs to be done after methods declaration.
+    __ACTIONS = {
+        OK: _on_ok,
+        UNCLEAR: _on_unclear,
+        ERROR: _on_error,
+        WARNING: _on_warning
+    }
+
 
 class MonoValidator(ValidatorFilter):
 
@@ -141,50 +141,3 @@ class MultiValidator(ValidatorFilter):
     # Should probs override _input_check_order(...)
 
     pass
-
-
-def make_check_date_between(date1: datetime, date2: datetime, inclusive: bool = False) -> Callable[[datetime, Mapping], Tuple[int, str]]:
-    '''
-    Parameters:
-        date1 : datetime
-            The first date
-        date2 : datetime
-            The second date
-        inclusive : bool = False
-            Whether to include the two dates in the accepted interval
-
-    Returns:
-        Callable[datetime] : Checks whether the given datetime is in-between the two given dates.
-        Will choose the smallest date as start date
-    '''
-    start = min([date1, date2])
-    end = max([date1, date2])
-    inclusive = inclusive
-
-    def check_date_between(d: datetime, state: Mapping = None) -> Tuple[bool, str]:
-        '''
-        Parameters:
-            d : datetime
-                A date to check
-            state : Mapping
-                State for this check, unused.
-        Returns:
-            `ValidatorFilter.OK` and None if parameter d is after start and before end.
-            `ValidatorFilter.ERROR` and a string error if it isn't. If inclusive is False also
-            does not consider `start` and `end` as "between".
-            Does NOT check timezones.
-        '''
-        if inclusive:
-            if start <= d and d <= end:
-                return ValidatorFilter.OK, None
-            else:
-                return ValidatorFilter.ERROR,
-                "Date not between {} and {} (inclusive).".format(start, end)
-        else:
-            if start < d and d < end:
-                return ValidatorFilter.OK, None
-            else:
-                return ValidatorFilter.ERROR,
-                "Date not between {} and {} (non-inclusive).".format(start, end)
-
-    return check_date_between
